+++ src/attacks/attack_extreme.c
@@ -7,7 +7,7 @@
 #include "../header_fields.h"
 #include "attack_extreme.h"
 
-bool attack_extreme(const char *output_filename, int index) {
+bool attack_extreme(const char *output_filename, uint8_t index) {
     bool is_header_tested = true;
 
     tar_archive archive;
