+++ src/attacks/attack_extreme.h
@@ -1,12 +1,14 @@
 #ifndef ATTACK_EXTREME_H
 #define ATTACK_EXTREME_H
 
+#include <stdint.h>
+
 /**
  * Create a TAR archive with an extreme value for different fields.
  * @param output_filename The name of the output file to generate.
  * @param index The index of the attack to perform.
  * @return True if a header has been tested, false otherwise.
  */
-bool attack_extreme(const char *output_filename, int index);
+bool attack_extreme(const char *output_filename, uint8_t index);
 
 #endif /* ATTACK_EXTREME_H */
