+++ src/attacks/attack_zero_size_file.c
@@ -5,7 +5,7 @@
 #include "../tar_header.h"
 #include "attack_zero_size_file.h"
 
-bool attack_zero_size_file(const char *output_filename, int index) {
+bool attack_zero_size_file(const char *output_filename, uint8_t index) {
     (void)index;
     tar_archive archive;
     init_tar_archive(&archive);
